# packages/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateKVNamespaceInput": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/CreateKVNamespaceInput.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "external_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["name"],
        "type": "object"
      },
...
      "KVNamespace": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/KVNamespace.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "external_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "oauth_client_id": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          }
        },
        "required": ["id", "name", "created_at"],
        "type": "object"
      },
      "KVNamespaceMetrics": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/KVNamespaceMetrics.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "available": {
            "type": "boolean"
          },
          "reads": {
            "format": "int64",
            "type": "integer"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          },
          "writes": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": ["available", "reads", "writes", "size"],
        "type": "object"
      },
...
      "OpenAPIError": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OpenAPIError.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": ["error"],
        "type": "object"
      },
...
      "UpdateKVNamespaceInput": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/UpdateKVNamespaceInput.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["name"],
        "type": "object"
      },
...
      "WorkerKVKey": {
        "additionalProperties": false,
        "properties": {
          "key": {
            "type": "string"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": ["key", "size"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT or personal access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Nanoflare Control API",
    "version": "v1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/organizations/{orgID}/kv-namespaces": {
      "get": {
        "operationId": "listKVNamespaces",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KVNamespace"
                  },
                  "type": ["array", "null"]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "listKVNamespaces",
        "tags": ["KV"]
      },
      "post": {
        "operationId": "createKVNamespace",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKVNamespaceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KVNamespace"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "createKVNamespace",
        "tags": ["KV"]
      }
    },
    "/v1/organizations/{orgID}/kv-namespaces/{namespaceID}": {
      "delete": {
        "operationId": "deleteKVNamespace",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "namespaceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "deleteKVNamespace",
        "tags": ["KV"]
      },
      "get": {
        "operationId": "getKVNamespace",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "namespaceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KVNamespace"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "getKVNamespace",
        "tags": ["KV"]
      },
      "patch": {
        "operationId": "updateKVNamespace",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "namespaceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKVNamespaceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KVNamespace"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "updateKVNamespace",
        "tags": ["KV"]
      }
    },
    "/v1/organizations/{orgID}/kv-namespaces/{namespaceID}/analytics": {
      "get": {
        "operationId": "getKVNamespaceMetrics",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "namespaceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KVNamespaceMetrics"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "getKVNamespaceMetrics",
        "tags": ["KV"]
      }
    },
...
    "/v1/organizations/{orgID}/workers/{workerID}/kv-namespaces/{namespaceID}/values": {
      "get": {
        "operationId": "listWorkerKV",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workerID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "namespaceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkerKVKey"
                  },
                  "type": ["array", "null"]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "listWorkerKV",
        "tags": ["KV"]
      }
    },
...
  }
}